feat: Add Turso database integration and UI improvements
- Add Turso database setup with charts, podcasts, episodes tables
- Add GitHub Actions workflow for PI data sync
- Add CodeRabbit config for AI code reviews
- Refactor player components for better modularity
- Add predictive back gesture support
- Update home screen loading patterns
- Add local.properties with Turso URL reference
