+++ .coderabbit.yaml
@@ -1,0 +1,21 @@
+# yaml-language-server: $schema=https://coderabbit.ai/integrations/schema.v2.json
+language: "en-US"
+tone_instructions: "Be concise and helpful"
+early_access: true
+reviews:
+  profile: "chill"
+  request_changes_workflow: false
+  high_level_summary: true
+  poem: false
+  review_status: true
+  collapse_walkthrough: true
+  path_instructions:
+    - path: "**/*.kt"
+      instructions: "Focus on Kotlin best practices, null safety, and Compose patterns"
+    - path: "**/*.ts"
+      instructions: "Focus on TypeScript best practices and Cloudflare Workers patterns"
+  auto_review:
+    enabled: true
+    drafts: false
+chat:
+  auto_reply: true
