+++ feature/home/src/main/java/cx/aswin/boxcast/feature/home/components/TopControlBar.kt
@@ -12,6 +12,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
@@ -95,6 +96,7 @@ fun TopControlBar(
         modifier = modifier
             .fillMaxWidth()
             .background(backgroundColor)
+            .statusBarsPadding()
             .padding(horizontal = 16.dp, vertical = verticalPadding),
         horizontalArrangement = Arrangement.SpaceBetween,
         verticalAlignment = Alignment.CenterVertically
