+++ feature/info/build.gradle.kts
@@ -46,4 +46,5 @@ dependencies {
     implementation(libs.androidx.material.icons.extended)
     implementation(libs.coil.compose)
     implementation(libs.kotlinx.coroutines.android)
+    implementation(libs.androidx.palette.ktx)
 }
