+++ feature/player/build.gradle.kts
@@ -50,4 +50,5 @@ dependencies {
     implementation(libs.androidx.material.icons.extended)
     implementation(libs.coil.compose)
     implementation(libs.kotlinx.coroutines.android)
+    implementation(libs.smooth.corner.rect)
 }
