+++ feature/player/src/main/java/cx/aswin/boxcast/feature/player/PlayerScreen.kt
@@ -346,12 +346,7 @@ fun PlayerContent(
     }
 }
 
-private fun formatTime(ms: Long): String {
-    val totalSeconds = ms / 1000
-    val minutes = totalSeconds / 60
-    val seconds = totalSeconds % 60
-    return String.format("%02d:%02d", minutes, seconds)
-}
+
 
 /**
  * M3 Expressive: Bouncy Audio Visualizer using Spring physics.
